Crate interledger_btp

interledger-btp

Client and server implementations of the Bilateral Transport Protocol (BTP). This is a WebSocket-based protocol for exchanging ILP packets between directly connected peers.

Because this protocol uses WebSockets, only one party needs to have a publicly-accessible HTTPS endpoint but both sides can send and receive ILP packets.

Structs

BtpOpenSignupAccount

BtpOutgoingService
A container for BTP/WebSocket connections that implements OutgoingService for sending outgoing ILP Prepare packets over one of the connected BTP connections.

BtpService

Traits

BtpAccount

BtpOpenSignupStore
The interface for Store implementations that allow open BTP signups. Every incoming WebSocket connection will automatically have a BtpOpenSignupAccount created and added to the store.

BtpStore
The interface for Store implementations that can be used with the BTP Server.

Functions

connect_client
Create a BtpOutgoingService wrapping BTP connections to the accounts specified. Calling handle_incoming with an IncomingService will turn the returned BtpOutgoingService into a bidirectional handler.

create_open_signup_server
Same as create_server but it returns a BTP server that will accept new connections and create account records on the fly.

create_server
Returns a BtpOutgoingService that wraps all BTP/WebSocket connections that come in on the given address. Calling handle_incoming with an IncomingService will turn the returned BtpOutgoingService into a bidirectional handler.

parse_btp_url